+++ config/jest-config-ibm-cloud-cognitive/index.js
@@ -51,7 +51,6 @@ module.exports = {
   reporters: ['default'],
   setupFiles: [require.resolve('./setup/setupFiles')],
   setupFilesAfterEnv: [require.resolve('./setup/setupFilesAfterEnv')],
-  snapshotSerializers: ['enzyme-to-json/serializer'],
   testEnvironment: 'jsdom',
   testMatch: [
     // '<rootDir>/**/__tests__/**/*.js?(x)',
@@ -74,9 +73,10 @@ module.exports = {
     'e2e',
     'templates',
     '/umd/',
+    '/react-dnd/',
   ],
   transformIgnorePatterns: [
-    '[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$',
+    `/node_modules/(?!react-dnd|dnd-core|@react-dnd)`,
     'ace-node\\.js',
   ],
   watchPathIgnorePatterns: [