+++ packages/ibm-products/src/components/Datagrid/Datagrid/addons/CustomizeColumns/DraggableItemsList.js
@@ -11,6 +11,7 @@ import { Checkbox } from '@carbon/react';
 import { isColumnVisible } from './common';
 import DraggableElement from '../../DraggableElement';
 import { pkg } from '../../../../../settings';
+import getColTitle from '../../../utils/getColTitle';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 
@@ -19,7 +20,6 @@ export const DraggableItemsList = ({
   filterString,
   focusIndex,
   getNextIndex,
-  isTableSortable,
   moveElement,
   onSelectColumn,
   setAriaRegionText,
@@ -31,36 +31,22 @@ export const DraggableItemsList = ({
       {columns
         // hide the columns without Header, e.g the sticky actions, spacer
         .filter((colDef) => {
-          const sortableTitle =
-            isTableSortable && colDef.Header().props.children.props?.title;
-          return (
-            (!!colDef.Header.props && !!colDef.Header.props?.title) ||
-            (isTableSortable && !!sortableTitle)
-          );
+          return !!getColTitle(colDef);
         })
+        .filter(Boolean)
         .filter((colDef) => !colDef.isAction)
         .filter((colDef) => {
-          const sortableTitle =
-            isTableSortable && colDef.Header().props.children.props?.title;
           return (
             filterString.length === 0 ||
-            ((isTableSortable
-              ? sortableTitle?.toLowerCase().includes(filterString)
-              : colDef.Header.props?.title
-                  ?.toLowerCase()
-                  .includes(filterString)) &&
+            (getColTitle(colDef)?.toLowerCase().includes(filterString) &&
               colDef.id !== 'spacer')
           );
         })
         .map((colDef, i) => {
-          const isSortableColumn = !!colDef.canSort && !!isTableSortable;
-          const sortableTitle =
-            isTableSortable && colDef.Header().props.children.props?.title;
+          const colHeaderTitle = getColTitle(colDef);
           const searchString = new RegExp('(' + filterString + ')');
           const res = filterString.length
-            ? isSortableColumn
-              ? sortableTitle.toLowerCase().split(searchString)
-              : colDef.Header.props.title.toLowerCase().split(searchString)
+            ? colHeaderTitle.toLowerCase().split(searchString)
             : null;
           const firstWord =
             res !== null
@@ -73,9 +59,7 @@ export const DraggableItemsList = ({
               ? res[0] === ''
                 ? `<strong>${firstWord}</strong>` + res[2]
                 : firstWord + `<strong>${res[1]}</strong>` + res[2]
-              : isSortableColumn
-              ? sortableTitle
-              : colDef.Header.props?.title;
+              : colHeaderTitle;
           const isFrozenColumn = !!colDef.sticky;
           const listContents = (
             <>
@@ -85,12 +69,8 @@ export const DraggableItemsList = ({
                 disabled={isFrozenColumn}
                 onChange={(_, { checked }) => onSelectColumn(colDef, checked)}
                 id={`${blockClass}__customization-column-${colDef.id}`}
-                labelText={
-                  isSortableColumn ? sortableTitle : colDef.Header.props?.title
-                }
-                title={
-                  isSortableColumn ? sortableTitle : colDef.Header.props?.title
-                }
+                labelText={colHeaderTitle}
+                title={colHeaderTitle}
                 className={`${blockClass}__customize-columns-checkbox`}
                 hideLabel
               />
@@ -112,9 +92,7 @@ export const DraggableItemsList = ({
               id={`dnd-datagrid-columns-${colDef.id}`}
               type="column-customization"
               disabled={filterString.length > 0 || isFrozenColumn}
-              ariaLabel={
-                isSortableColumn ? sortableTitle : colDef.Header.props?.title
-              }
+              ariaLabel={colHeaderTitle}
               onGrab={setAriaRegionText}
               isFocused={focusIndex === i}
               moveElement={moveElement}
@@ -149,7 +127,6 @@ DraggableItemsList.propTypes = {
   filterString: PropTypes.string.isRequired,
   focusIndex: PropTypes.number.isRequired,
   getNextIndex: PropTypes.func.isRequired,
-  isTableSortable: PropTypes.bool,
   moveElement: PropTypes.func.isRequired,
   onSelectColumn: PropTypes.func.isRequired,
   setAriaRegionText: PropTypes.func.isRequired,